requirements.txt
Babel
dash
dash_bootstrap_components
python-i18n[YAML]
pandas
plotly
python_i18n
pip
wheel
Brotli
click
Flask
Flask-Compress
future
itsdangerous
Jinja2
MarkupSafe
numpy
python-dateutil
pytz
retrying
setuptools
six
Werkzeug
gunicorn